refactor(pages/admin): AdminPage page 레이어로 분리
