+++ src/advanced/__tests__/advanced.test.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { describe, expect, test } from "vitest";
 import { act, fireEvent, render, screen, within } from "@testing-library/react";
 import { CartPage } from "../../refactoring/components/CartPage";
-import { AdminPage } from "../../refactoring/components/AdminPage";
+import AdminPage from "@/refactoring/pages/admin";
 import { Coupon } from "../../types";
 import {
   Product,
